# perf: don't collect more info than needed when resolving jest functions