@SimenB 2022-11-04
Woah, surprised by the big difference! Great job 👍