@G-Rath 2022-11-04
I just tried the same optimization to `scopeHasLocalReference` but it doesn't result in any noticeable win so am going to land this as-is; that makes sense anyway because what we're doing there actually requires us most of the time to recuse through everything, and it's only used by two rules 🤷