@SLKnutson 2022-11-04
@G-Rath 

It might be possible to improve this even more. There is a `set` field on Scope that I think has a map from variable name => variable, so each level check can use that instead of iterating the variables. When I swapped it out for my project, I got the same result as the existing implementation, but I'm not 100% sure that it works in all scenarios. If it can be used, the performance should be even better. Also, it can probably be used in `scopeHasLocalReference` too.
Thanks!


```
const resolveScope = (scope, identifier) => {
  let currentScope = scope;
  while (currentScope !== null) {
    const ref = currentScope.set.get(identifier);
    if (ref && ref.defs.length !== 0) {
      const importDetails = describePossibleImportDef(ref.defs[ref.defs.length - 1]);
      return importDetails?.local === identifier ?  importDetails : 'local';
    }
    currentScope = currentScope.upper;
  }
  return null;
};
```