@G-Rath 2022-11-04
btw @SLKnutson if you wouldn't mind, I would be interested in seeing what the difference is if you are/are not using `@jest/globals`, since that should trigger slightly different code paths that I think would make a difference. (i.e. if your project is using `@jest/globals`, switching to using the native globals, and vice versa)