@SLKnutson 2022-11-05
@G-Rath 
I haven't tried the global stuff yet, but with the change I mentioned above to `resolveScope`, I got these speed-ups before/after. Seems like it cuts most rule run times by a factor of 2 or 3, which is noticeable for bigger projects. Thanks!

Rule                                                       | Time (ms) | Relative
:----------------------------------------------------------|----------:|--------:
jest/consistent-test-it                                    |  2892.622 |     2.6%
jest/no-disabled-tests                                     |  1116.793 |     1.0%
jest/no-identical-title                                    |  1051.963 |     1.0%
jest/no-duplicate-hooks                                    |  1045.640 |     1.0%
jest/require-top-level-describe                            |  1013.668 |     0.9%
jest/no-alias-methods                                      |   599.548 |     0.5%
jest/prefer-hooks-on-top                                   |   558.858 |     0.5%
jest/valid-title                                           |   549.298 |     0.5%
jest/valid-expect                                          |   540.753 |     0.5%
jest/prefer-comparison-matcher                             |   537.186 |     0.5%
jest/prefer-to-contain                                     |   536.460 |     0.5%
jest/prefer-todo                                           |   533.691 |     0.5%
jest/no-export                                             |   531.140 |     0.5%
jest/prefer-equality-matcher                               |   522.748 |     0.5%
jest/no-focused-tests                                      |   518.938 |     0.5%
jest/no-large-snapshots                                    |   518.034 |     0.5%
jest/require-to-throw-message                              |   514.279 |     0.5%
jest/prefer-to-have-length                                 |   513.314 |     0.5%
jest/valid-describe-callback                               |   512.982 |     0.5%
jest/no-test-prefixes                                      |   511.949 |     0.5%
jest/prefer-strict-equal                                   |   508.050 |     0.5%
jest/no-interpolation-in-snapshots                         |   507.236 |     0.5%
jest/prefer-expect-resolves                                |   503.346 |     0.5%
jest/no-deprecated-functions                               |   268.651 |     0.2%
jest/no-jasmine-globals                                    |   268.419 |     0.2%
jest/max-nested-describe                                   |   110.651 |     0.1%
jest/no-mocks-import                                       |    36.196 |     0.0%
jest/prefer-spy-on                                         |    20.564 |     0.0%
jest/no-commented-out-tests                                |     9.545 |     0.0%


Rule                                                       | Time (ms) | Relative
:----------------------------------------------------------|----------:|--------:
jest/consistent-test-it                                    |  1511.273 |     1.5%
jest/no-disabled-tests                                     |   362.679 |     0.4%
jest/no-identical-title                                    |   331.576 |     0.3%
jest/no-duplicate-hooks                                    |   308.397 |     0.3%
jest/require-top-level-describe                            |   291.826 |     0.3%
jest/no-deprecated-functions                               |   267.655 |     0.3%
jest/no-jasmine-globals                                    |   265.734 |     0.3%
jest/no-alias-methods                                      |   215.248 |     0.2%
jest/valid-title                                           |   194.013 |     0.2%
jest/prefer-hooks-on-top                                   |   179.216 |     0.2%
jest/valid-expect                                          |   178.825 |     0.2%
jest/prefer-comparison-matcher                             |   176.338 |     0.2%
jest/prefer-todo                                           |   167.468 |     0.2%
jest/prefer-to-contain                                     |   167.344 |     0.2%
jest/prefer-expect-resolves                                |   158.118 |     0.2%
jest/no-export                                             |   157.577 |     0.2%
jest/no-interpolation-in-snapshots                         |   157.258 |     0.2%
jest/no-large-snapshots                                    |   156.854 |     0.2%
jest/prefer-strict-equal                                   |   152.639 |     0.2%
jest/prefer-equality-matcher                               |   151.352 |     0.2%
jest/valid-describe-callback                               |   150.444 |     0.1%
jest/require-to-throw-message                              |   148.396 |     0.1%
jest/no-focused-tests                                      |   147.469 |     0.1%
jest/no-test-prefixes                                      |   147.116 |     0.1%
jest/prefer-to-have-length                                 |   143.708 |     0.1%
jest/max-nested-describe                                   |    91.314 |     0.1%
jest/no-mocks-import                                       |    38.051 |     0.0%
jest/prefer-spy-on                                         |    20.127 |     0.0%
jest/no-commented-out-tests                                |     9.996 |     0.0%